# [WIP] Improve read performance from memcache