vs goon

```
# in $GOPATH/src/go.mercari.io/datastore/benchmark/vs_goon
$ goapp test -bench . -benchmem -benchtime 10s
BenchmarkGoonFromAEMemcache-8             1000      15112614 ns/op     1591120 B/op       37224 allocs/op
BenchmarkBoomFromAEMemcache-8              300      33531617 ns/op     6717238 B/op      147126 allocs/op

$ goapp test -bench . -benchmem -benchtime 20s
BenchmarkGoonFromAEMemcache-8             2000      13585816 ns/op     1586297 B/op       37119 allocs/op
BenchmarkBoomFromAEMemcache-8             1000      28089797 ns/op     6717331 B/op      147125 allocs/op
```

datastore wrapper is slower than goon x2.2.
goon is caching entity by struct, but datastore wrapper is caching entity by datastore.PropertyList.
datastore.PropertyList will be mapping to struct. It using reflect package. Maybe it is too late.

1. Is this a problem?
2. Can we solve this issue?