@vvakame 2017-12-15
```
# in $GOPATH/src/go.mercari.io/datastore/benchmark/vs_goon
$ goapp test -bench . -benchmem -benchtime 10s
BenchmarkGoonFromAEMemcache-8     500	  20461281 ns/op	 1601833 B/op	   37434 allocs/op
BenchmarkBoomFromAEMemcache-8     300	  33754387 ns/op	 6717086 B/op	  147125 allocs/op
BenchmarkGoonFromLocalCache-8   10000	   1117324 ns/op	  372772 B/op	   12908 allocs/op
BenchmarkBoomFromLocalCache-8    2000	   5524116 ns/op	 1429887 B/op	   38461 allocs/op
```

🤔